[SES-130] Code improvement on transparency report
