+++ pages/core-unit/[code]/finances/reports/index.tsx
@@ -1,15 +1,19 @@
+import request from 'graphql-request';
 import { GetServerSidePropsContext } from 'next';
 import React from 'react';
+import { GRAPHQL_ENDPOINT } from '../../../../../src/config/endpoints';
 import { CoreUnitDto } from '../../../../../src/core/models/dto/core-unit.dto';
-import { fetchCoreUnits, SummarizedCoreUnit } from '../../../../../src/stories/components/core-unit-summary/core-unit-summary.mvvm';
+import { fetchCoreUnits } from '../../../../../src/stories/components/core-unit-summary/core-unit-summary.mvvm';
 import { TransparencyReport } from '../../../../../src/stories/containers/transparency-report/transparency-report';
-import { useTransparencyReportViewModel } from '../../../../../src/stories/containers/transparency-report/transparency-report.mvvm';
+import { CORE_UNIT_REQUEST } from '../../../../../src/stories/containers/transparency-report/transparency-report.api';
 
 export const getServerSideProps = async(context: GetServerSidePropsContext) => {
   const { query } = context;
   const code = query.code as string;
-  const [{ data }, coreUnits] = await Promise.all([
-    useTransparencyReportViewModel(code),
+  const { query: gqlQuery, filter } = CORE_UNIT_REQUEST(code);
+
+  const [data, coreUnits] = await Promise.all([
+    request(GRAPHQL_ENDPOINT, gqlQuery, filter),
     fetchCoreUnits()
   ]);
 
@@ -28,7 +32,7 @@ export const getServerSideProps = async(context: GetServerSidePropsContext) => {
 };
 
 interface TransparencyProps {
-  coreUnits: SummarizedCoreUnit[],
+  coreUnits: CoreUnitDto[],
   cu: CoreUnitDto
 }
